wip building up acquisition class, cleanup rx.py
