+++ rx.py
@@ -39,7 +39,7 @@ import argparse
 import numpy as np
 from matplotlib import pyplot as plt
 import torch
-from radae import RADAE,complex_bpf
+from radae import RADAE,complex_bpf,acquisition
 
 parser = argparse.ArgumentParser()
 
@@ -111,73 +111,6 @@ if args.plots:
    ax[1].axis([0,len(rx)/model.Fs,0,3000])
    ax[1].set_title('After BPF')
 
-class acquisition():
-   def __init__(self,Fs,Rs,M,Nmf,p,frange=100,fstep=2.5,Pacq_error = 0.0001):
-      self.Fs = Fs
-      self.Rs = Rs
-      self.M = M
-      self.Nmf = Nmf
-      self.p = p
-      self.frange = frange
-      self.fstep=fstep
-      self.Pacq_error = Pacq_error
-      self.fcoarse_range = np.arange(-frange/2,frange/2,fstep)
-
-      # pre-calculate to speeds things up a bit
-      p_w = np.zeros((len(self.fcoarse_range), M), dtype=np.csingle)
-      f_ind = 0
-      for f in self.fcoarse_range:
-         w = 2*np.pi*f/Fs
-         p_w[f_ind,] = np.exp(1j*w*np.arange(M)) * p
-
-         f_ind + f_ind + 1
-      self.p_w = p_w
-
-   def detect_pilots(self, rx):
-      M = self.M
-      Nmf = self.Nmf
-
-      # We need a buffer of two and bit modem frames, and search of one Nmf, TODO to reduce
-      # latency this could be reduced to a one symbol (M sample) search and Nmf+2*M sample buffer
-      assert len(rx) == self.Nmf*2+M
-
-      Dt1 = np.zeros((self.Nmf,len(self.fcoarse_range)), dtype=np.csingle)
-      Dt2 = np.zeros((self.Nmf,len(self.fcoarse_range)), dtype=np.csingle)
-      Dtmax12 = 0
-      tmax = 0
-      fmax = 0
-
-      # Search modem frame for maxima in correlation between pilots and received signal, over
-      # a grid of time and frequency steps.  Note we only correlate on the M samples after the
-      # cyclic prefix, so tmax will be Ncp samples after the start of the modem frame
-
-      for t in range(Nmf):
-         f_ind = 0
-         for f in self.fcoarse_range:
-            w = 2*np.pi*f/Fs
-            w_vec = np.exp(-1j*w*np.arange(M))
-            Dt1[t,f_ind] = np.dot(np.conj(w_vec*rx[t:t+M]),p)
-            Dt2[t,f_ind] = np.dot(np.conj(w_vec*rx[t+Nmf:t+Nmf+M]),p)
-            Dt12 = np.abs(Dt1[t,f_ind]) + np.abs(Dt2[t,f_ind])
-            if Dt12 > Dtmax12:
-               Dtmax12 = Dt12
-               tmax = t
-               fmax = f 
-               f_ind_max =  f_ind
-            f_ind = f_ind + 1
-
-      # Ref: radae.pdf "Pilot Detection over Multiple Frames"
-      sigma_r1 = np.mean(np.abs(Dt1))/((np.pi/2)**0.5)
-      sigma_r2 = np.mean(np.abs(Dt2))/((np.pi/2)**0.5)
-      sigma_r = (sigma_r1 + sigma_r2)/2.0
-      Dthresh = 2*sigma_r*np.sqrt(-np.log(self.Pacq_error/5.0))
-
-      candidate = False
-      if Dtmax12 > Dthresh:
-         candidate = True
-     
-      self.Dt1 = Dt1
-      return candidate, Dthresh, Dtmax12, tmax, fmax, f_ind_max 
 
 # Acquisition - 1 sample resolution timing, coarse/fine freq offset estimation
 
@@ -189,12 +122,7 @@ if args.pilots:
    Rs = model.Rs
 
    acq = acquisition(Fs,Rs,M,Nmf,p)
-   """
-   # correlation at various time and freq offsets
-   fcoarse_range = np.arange(-frange/2,frange/2,fstep)
-   Dt1 = np.zeros((Nmf,len(fcoarse_range)), dtype=np.csingle)
-   Dt2 = np.zeros((Nmf,len(fcoarse_range)), dtype=np.csingle)
-   """
+ 
    # optional acq_test variables 
    tmax_candidate_target = Ncp + Ntap/2
    acq_pass = 0
@@ -203,66 +131,16 @@ if args.pilots:
    tmax_candidate = 0 
    acquired = False
    state = "search"
-
-   """
-   Pacq_error = 0.0001
-   # pre-calculate to speeds things up a bit
-   p_w = np.zeros((len(fcoarse_range), M), dtype=np.csingle)
-   f_ind = 0
-   for f in fcoarse_range:
-      w = 2*np.pi*f/Fs
-      p_w[f_ind,] = np.exp(1j*w*np.arange(M)) * p
-
-      f_ind + f_ind + 1
-   """
-
    mf = 1
-   """
+   
    if len(args.write_Dt):
-      print(Dt1.shape)
       fD=open(args.write_Dt,'wb')
-   """
 
    while not acquired and len(rx) >= 2*Nmf+M:
-      """
-      # Search modem frame for maxima in correlation between pilots and received signal, over
-      # a grid of time and frequency steps.  Note we only correlate on the M samples after the
-      # cyclic prefix, so tmax will be Ncp samples after the start of the modem frame
-      Dtmax12 = 0
-      tmax = 0
-      fmax = 0
-
-      for t in range(Nmf):
-         f_ind = 0
-         for f in fcoarse_range:
-            #D[t,f_ind] = np.dot(np.conj(rx[t:t+M]),p_w[f_ind,:])
-            w = 2*np.pi*f/Fs
-            w_vec = np.exp(-1j*w*np.arange(M))
-            Dt1[t,f_ind] = np.dot(np.conj(w_vec*rx[t:t+M]),p)
-            Dt2[t,f_ind] = np.dot(np.conj(w_vec*rx[t+Nmf:t+Nmf+M]),p)
-            Dt12 = np.abs(Dt1[t,f_ind]) + np.abs(Dt2[t,f_ind])
-            if Dt12 > Dtmax12:
-               Dtmax12 = Dt12
-               tmax = t
-               fmax = f 
-               f_ind_max =  f_ind
-            f_ind = f_ind + 1
-      """
       candidate, Dthresh, Dtmax12, tmax, fmax, f_ind_max = acq.detect_pilots(rx[:2*Nmf+M])
-      """
       if len(args.write_Dt):
-         Dt1.tofile(fD)
+         acq.Dt1.tofile(fD)
       
-      # Ref: radae.pdf "Pilot Detection over Multiple Frames"
-      sigma_r1 = np.mean(np.abs(Dt1))/((np.pi/2)**0.5)
-      sigma_r2 = np.mean(np.abs(Dt2))/((np.pi/2)**0.5)
-      sigma_r = (sigma_r1 + sigma_r2)/2.0
-      Dthresh = 2*sigma_r*np.sqrt(-np.log(Pacq_error/5.0))
-
-      candidate = False
-      if Dtmax12 > Dthresh:
-         candidate = True
-      """
       # post process with a state machine that looks for 3 consecutive matches with about the same tmining offset      
       print(f"{mf:2d} state: {state:10s} Dthresh: {Dthresh:5.2f} Dtmax12: {Dtmax12:5.2f} tmax: {tmax:4d} tmax_candidate: {tmax_candidate:4d} fmax: {fmax:6.2f}")
